requirements.txt
html5lib==0.999
httplib2==0.8
psycopg2==2.7.1
pycrypto==2.6.1
python-dateutil==1.5
pytz==2012c
tabula-py==0.4.1
urllib3==1.7.1
Flask
Jinja2==2.8.1
gunicorn
